+++ .travis.yml
@@ -1,0 +1,6 @@
+language: node_js
+node_js:
+  - v10
+  - v8
+  - v6
+  - v4
