--- PULL_REQUEST_TEMPLATE.md
@@ -1,17 +1,0 @@
-### This is a
-
-- [ ] Breaking change
-- [ ] New feature
-- [ ] Bugfix
-
-### I have
-
-- [ ] Tested my changes by generating a new component
-- [ ] Updated [`CHANGELOG.md`](CHANGELOG.md)
-- [ ] Updated [`README.md`](README.md)
-
-### What's changed
-
-> Add a quick dot-point outline 
-> Add instructions on testing
-> Add JIRA or Trello etc link
