+++ __tests__/app.js
@@ -1,0 +1,16 @@
+'use strict';
+const path = require('path');
+const assert = require('yeoman-assert');
+const helpers = require('yeoman-test');
+
+describe('generator-td-react-boilerplate:app', () => {
+  beforeAll(() => {
+    return helpers
+      .run(path.join(__dirname, '../generators/app'))
+      .withPrompts({ someAnswer: true });
+  });
+
+  it('creates files', () => {
+    assert.file(['dummyfile.txt']);
+  });
+});
