+++ generators/app/templates/.babelrc
