+++ generators/app/templates/.editorconfig
