+++ generators/app/templates/.gitignore
