+++ generators/app/templates/public/images/bg.jpg
