+++ generators/app/templates/public/images/favicon.png
