+++ generators/app/templates/public/images/loader.gif
