+++ generators/app/templates/public/index.html
