+++ generators/app/templates/server/server.js
