+++ generators/app/templates/src/js/app.js
