+++ generators/app/templates/src/js/components/landing.js
