+++ generators/app/templates/src/scss/general/_general.scss
