+++ generators/app/templates/src/scss/general/_settings.scss
