+++ generators/app/templates/src/scss/styles.scss
