+++ generators/app/templates/test/components/landing.js
