+++ generators/app/templates/test/globals.js
