+++ generators/app/templates/test/mocha.opts
