+++ generators/app/templates/webpack.config.js
