--- util/pascalcase.js
@@ -1,6 +1,0 @@
-const { flow, camelCase, upperFirst } = require('lodash');
-
-module.exports = flow(
-  camelCase,
-  upperFirst,
-);
